+++ packages/backend-graphql/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/backend-graphql
 
+## 0.1.1-alpha.2
+
+### Patch Changes
+
+- b2b0c2d: force version bump
+- Updated dependencies [b2b0c2d]
+- Updated dependencies [395c8f0]
+  - @aws-amplify/backend-output-schemas@0.1.1-alpha.1
+  - @aws-amplify/plugin-types@0.1.1-alpha.2
+
 ## 0.1.1-alpha.1
 
 ### Patch Changes