+++ packages/create-amplify/CHANGELOG.md
@@ -1,5 +1,18 @@
 # create-amplify
 
+## 0.2.0-alpha.4
+
+### Patch Changes
+
+- b2b0c2d: force version bump
+- 395c8f0: Add identityPoolId to output, and set delete policy on user pool to delete.
+
+## 0.2.0-alpha.3
+
+### Patch Changes
+
+- 9a1cf73: Execute npm init with --yes flag and split dependencies and dev dependencies when installing packages into the project
+
 ## 0.2.0-alpha.2
 
 ### Patch Changes