+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/sandbox
 
+## 0.1.1-alpha.3
+
+### Patch Changes
+
+- b2b0c2d: force version bump
+- Updated dependencies [b2b0c2d]
+- Updated dependencies [395c8f0]
+  - @aws-amplify/backend-deployer@0.1.1-alpha.2
+  - @aws-amplify/client-config@0.1.1-alpha.3
+
 ## 0.1.1-alpha.2
 
 ### Patch Changes